requirements.txt
gitpython>=3.1.31
requests>=2.31.0